Define the structures and APIs for using mappings to determine good package names
The 'defaultPackageName' in a mapping file defines the package to be used for
all classes generated with the same base URI.

At the moment the lookup is not implemented, so it will always fall back to the
value defined via the DEFAULT_PACKAGE_NAME feature. Note that the feature URI changed
to match the move from PojoGenerator to Generator.