Use the canonical name for the ClassName rawClassName for existing classes
The canonical name is the one used for 'import' statements, which is what we
typically do with ClassName instances. Unfortunately we also need to be able
to query the class loader when looking up existing non-primitive classes, so
that method now needs to attempt some trial-and-error with the name.

The code is inspired by common-lang's ClassUtils#getClass(), although that
one doesn't handle inner-inner classes from what I can see.