Allow to generate a different class than what is used for referencing it
In some cases the generation should create a base class, and the code then extends
this further with additional functionality. In those cases we want code _using_ the type
refer to the extended class, but the generator should only produce the base class.

This can be done now by adding a "generatedClassName", which defaults to the value
of "className".