Generate a #parse() method for enums
The #parse() method takes the string value, and tries to match it against a known value. If
no such value is known an IllegalArgumentException is thrown.

For old-school enums this also adds a #values() method, which currently produces a List<TYPE>
of all values. Eventually this will change to be compatible with regular enums by producing a
TYPE[], but our JavaWriter/ClassName code cannot produce arrays yet.