Allow configuring the enumStyle per mapping
The value from the feature is now used as a fallback/default value, each mapping
can override that by setting 'enumStyle' to 'class' or 'enum'.

Note that using other values (any Kind is allowed) will likely lead to exceptions
when generating the code due to the missing generator of that kind.